65. Controlling which properties can be changed (no username change)(Section 9: API Platform - Authorization, Data Validation, Serialization)
In this lecture we will enable that put operation for users.

But we have some exceptions so first off we would like to only allow users to modify their own profile.

And also we would like them to be able to modify every possible property.

So usually systems do not allow users to change their username so we would like to allow users to modify

certain properties of their profile but not username.

So first modification we can do to the access control is seems we do not have get out or we only compare

object to the user because a user is the variable as they're currently authenticated users are symphony

provides that.

So whoever you are logged in as will be available under user and we are inside the user anti-Tea itself.

So we can use this object variable to relate to the user which ID will provide in the user L next.

We have used groups for the normalization which is always the process of serializing the data or sending

the data in adjacent format back to the client and we have assigned a read group to all the operations

that will return some data back and we use the read group.

Now lets rename that to get because we all have different groups and different cases where we would

like to return data back in different format.

So now this normalization context is defined for every operation.

So even if you do a post or put Operation you got some data back.

So you will get the data of the modified resource and we would like to be able to control what is being

returned.

And also we will create a different setting which is not normalization of group but then normalization

of group so we can move this normalization context setting into the operation.

So now in this case we will only set the normalization group for the Get operation so all the other

operations will be serialized using the default group.

So every property that basically exists would be serialized.

Now let's copy that to the pute operation.

But this time we will call it our then normalization context and simply put that then normalization

context is we say what properties we allowed to send when the client is actually sending to us.

So the Delmon ization context with its groups in this case we want our pute group for the operation.

So for each time user performs a boot operation we'll be able to specify using the groups which properties

exactly the user can even sense to us.

So all the properties on your entity that would not have put a group would be just ignored.

So they will never get to the system theyll be filtered out.

Also let's do the same for the post-operation.

So let's define the normalization group for the post-operation and get a group called post.

Now by adding this groups to the groups annotation on each individual property we allow to tell the

sinfulness Serializer component which properties we would like to get into each particular group.

So as we have to get group for sending data back we always want to send back a username and we only

want to accept user name coming from the client when it is up.

Post request and that's because we do not want their users to be able to change their username.

So if it would not have put a group it would be just ignored.

But it's a different case with the name.

So we always want users to provide their name when they're registering but we also allow them to modify

the user name when they they are modifying their profile.

Now every operation has two context.

So the first context is the normalization.

When you send the data back and then normalization is when you accept some data from the user.

So email did not have a group and we decided that for every operation that sends data back we want to

use the read group.

Now we changed that to the get group.

So e-mail didn't have any group and was never used.

When the process was normalization So when that data was returned back.

But now we want users to be able to send an e-mail to us and modify it.

And since we decided that put and post will have their own the normalization groups will just post and

post respectively.

We will have to add the group annotation to e-mail so it won't be filtered out.

So how it works.

Once again also mine to fix those definitions because when the normalization or normalization context

is inside an individual operation it has to be surrounded by curly braces and the name should be underscore

instead of the uppercase letter.

OK so we are done with setting everything up.

But before we test that please follow me to the next lecture when I will try to explain visually more

about the normalization the normalization context about the serialization groups and how we can achieve

what we want using them and actually how powerful that this.