66.1. Hasing password on User changes (PUT)(Section 9: API Platform - Authorization, Data Validation, Serialization)
So in the last lecture we talk about how we can control what data we can accept for certain operations

or all operations and what data should be send back generally or on a particular operation.

So now let's see how our current configuration of the user entity works when we create a new user.

So you can see that we have a problem that we apparently fixed before.

So when we create a user the data returned back as the new entity contains a password assault and every

other data that we want to hide.

And this is because we previously defined that for every operation that returns data.

The group should be get.

And then we added the get group only to the properties that we wanted to be returned.

And since we added this setting only to that get operation now we just let every other operation that

returns date.

I use that default group and the default group includes every proper team.

So now to actually fix that we'll have to add this normalization group either as our general setting

as we did before.

So now you can see that our data returned from the post operation did not contain the passport or any

other sensitive data.

And that's because we said that for every operation that will return any data we would like to use the

group get.