66.2. Hasing password on User changes (PUT)(Section 9: API Platform - Authorization, Data Validation, Serialization)
But we can also define this normalization context for all the other operations individually in case

you would like to return something different.

So we have added the normalization context to both post and put operations.

And now when testing it or we have to use another username.

Now when testing it you can see that we do not have those sensitive information return back because

we said that for every individual operation we would like to use that get group.

So if you do not need to send any different data depending on the operation you can just use the generic

setting the normalization context that is inside that API resource.

But if you would like to control what you return from every individual operation you can add a normalization

context inside every individual operation and then invent some new group so he doesn't have to be the

get group.

You can create another group blog like get from post and then add this specific group only to the properties

of the entity that you would like to control how it would be returned back or normalized.

Now let's try to create a new user and see if we will be able to modify this user username so you can

call it user to modify.

Well seems we already created that user so choose something unique.

OK now we have this user ID is forty seven.

It might be different in your environment because it depends how many users you are.

But now let's try to modify that user with the idea that is being returned from the post-operation And

let's see what group do we have on the username so we only have get and post so scenes for that put

operation we define that the normalization context is poort if we going to send the username it should

be ignored.

But first before we will be able to test that we need to log in as that new user that we just created.

So go to the loggin check tab and send the username and password to get a fresh token and update the

token inside authorization tap because that's that Doc and it's probably from last logging and now try

to modify the username maybe to the nasty username.

And let's see what happened.

So you can see that the user to modify is still this users username so the user name property was ignored.

If you will try to change the name it actually did work because it was allowed the name property was

allowed to be sent under the port operation for that we defined as our normalization group for the pute

operation.

Now let's try to modify the password of the user and send it the correct way so the password is matching

the retyped password and it seems that something did happen. Password is properly hashed.

But this brings us to another problem that when we will modify a password the token that we have generated

should be actually invalid because the password change.

So users should be forced to log in because they still have a token for the last possible that they

had.

But we will solve that a little bit later.