67. Comment resource operations (POST/PUT)(Section 9: API Platform - Authorization, Data Validation, Serialization)
So now that we allow our blog posts to be posted it's time to also allow the users back to create comments.

But as you will see in this lecture we will have a couple of problems with posting the comments and

as you might see on the list of those comments each comment has an outer but it also has to have a link

to an individual blog post so we all need a way to specify the blog post and also as it currently stands.

We have this outward anti-Tea subscriber which only assigns that out or as it currently authenticated

to user to a blog post.

But first let's copy the API Rhys-Jones definition from the blog post straight into the comment entity

us we'll be using basically the same set up for time being so we have item operations and collection

operations and we will add the put and post to those and we also would like to verify that the user

has to be authenticated to create a comment and that the user who would like to modify our particular

comment is the actual user that created that comment in the first place.

CommentPersister class assure us that the currently logged in user will be automatically assigned to the currently created comment.

And by the way we will see what will happen if even if our rissoles is unprotected we somehow have token

included and it is expired.

So that was actually this case that I used this token previously and even vote to register.

You do not need to provide a token because we only need a token when we explicitly tell symphonic that

it should have some access control and it gives us an error because we send our token that was expired

so now that we created this account and we have logged in as this new user we can use stop talkin and

see if adding comment will work.

So we are doing that pre-view request to update our headers so we can see that the token is being sent

inside outer zation header and now we all need to write some Jason documents to create new Calment So

the fields that comment has is definitely the content you need to provide some content for this comment.

Another property is the published date.

And since we are not automatically setting the publish date as we do have it out or we will do that

later in the next section.

But for now we have to manually type the publish date.

So let's see what happened.

We have five hundred arrowed that's also because we do not have any validation constraints.

And you can see that the message out on my tickly assigned an outer ID but the last column that block

post ID which is required is no.

And we will handle that in the next section of how we can send the relations when they are required.

So we will know what we should provide to assign this command to our particular block post.