72.1. Controlling how deep relations are serialized(Section 9: API Platform - Authorization, Data Validation, Serialization)
In the last lecture we have added soob resource and it was comments which you could see a list of comments

all the comments for our particular blog post.

Now let's see how that looks.

So we have the content of the comment but it might also be useful to display username.

So when you will display a list of comments you can say cause they're out of that comment.

And now we only have you we're I of that user so that doesn't really look viable to call to make like

10 requests for each individual user profile to fetch when you are displaying a list of comments.

And we can solve that by specifying groups for their normalization context and we will also see the

possible issues that we might have of it.

So to define specific normalization context for this sub resource we actually need to use the root name

of that sub resource.

So we only specify the verb for item operations or collection operations for those five generic actions

like when you fetch a list of things or you use the post or port.

But in this case these are the new route and it's different thing that you have a collection operation

of get on the comments.

So we need to use that route as the operation name.

And now to specify how that data would be normalized we can copy the normalization context setting from

another class and we will set our group to it.

So by default if you do not set the group for the anonymization context it will just use that default

and it will serialize all the fields.

So it doesn't matter if you have group specified on all the properties but now that we specified that

we would like to have a get group in the case where we are displaying a list of comments as a sub resource

under different URL.

So only comments for a particular blog posts and we have specified to use the group get.

We will have to obviously add that group to all the properties that we would like to be serialized.

First see what happens when we haven't added that group to the properties.

Our sub resource displays a list of comments but only the metal properties of it and everything else

is not there and that is because we haven't added the group's notation for individual properties of

the comment with the group name that we specified to be used in the normalization context so we chose

to use Choat choose to use that get group.

So now you will have to add a group get under all of the properties that you would like to be serialized.

So first let's add it to the content because we obviously would like the content to be present.

We also made the ID.

So LEDs are the group's annotation to ID and now after adding it you can just copy to published you

can copy it to the outer loop because we also would like that our done.

But in this particular case we do not want to added to the blog post because we already know the id

of the blog post.

We do not want to show it.

Now lets see what happened when we use the get group for this operation.

So that group used for normalization context is used for older entities so its not unique across one

entity.

If you said that you would like to use to get group for serialising a list of comments in this sub resource

operation mind that the user also has an get group on particular properties.

For example in their comments.

So if you are serializing a comment and comment has a list of users but user itself has a list of comments

it will just cause an infinite loop.

So each time you serialize a comment its sorry a user and user has a list of comments and its keep getting

back and back.