72.2. Controlling how deep relations are serialized(Section 9: API Platform - Authorization, Data Validation, Serialization)
So we will have to define a different group and that will be unique across all those entities and we

can call it get comments with out or so we will out this group for our operation.

So our submarine operation will have a normalization context a group of get command with our tour.

So lets modify everyplace in inside comment that we have to get group but also we all have to add this

group to that user entity but only in the places where we are only on the properties that we would like

to be included under every single comment.

So we might want that username so we can show it on the list of comments on our front and we definitely

would like the name like the full name of the user.

Well you might want to show it in the email but probably not.

If we are not showing it in user profile we should not also added.

So we are not adding it to the posts of the user the comments because we already have that information

and we do not want this infinite loops.

And now see what happens.

So if we have a list of comments for our particular blog post we have the content of that comments but

we also manage to serialized out our data particularly the username and the full name of that outer

of that comment.

So now when we will be working on our front end using one request we can fetch a list of all the comments

for a blog post.

And the user outer of and the data of the outer of that comment.

If you fetch a single comment where you haven't defined this normalization group you'll see that we

still have are you or I for that outer.

And also if you list all the comments inside your system you can see that it still is adding the blog

post and it is adding that out outer as you or I not including the full data of outer of that particular

comment.

So that's why we can use normalization groups or even Then harmonisation groups depending on what is

our use case.

So whatever there might be points where you want to include some data and sometimes you do not want

to include it.

For example for performance so that's how we can achieve it.