73. Embedding Author resource inside BlogPost(Section 9: API Platform - Authorization, Data Validation, Serialization)
Last time we have seen how we can include all the data out of a particular comment when we are serializing

comments for our particular blog post and only in that case if you will see a blog post it might be

also useful to include the outer name.

So each time you go into block post on your front end you would like to show who is that out of the

blog post and we can achieve it.

Similar way.

So let's copy that configuration for normalization context that we put inside the sub resource operation

and will do the same this time for the get item operation on the block posts.

So this is an operation where when you display a single lock post using its ID.

Now we use to group get comment with our door.

But since we might want different data serialized Not necessarily but in case you want to do that and

to avoid any possible infinite loops that we've seen before which probably should would happen this

time.

If for example we'll serialize comments then we have to define these groups on individual properties

so we want this to be unique.

That's why for every possible scenario when we would like to serialize feelings in a specialized manner.

So you will probably have to always find a new group name for that.

Now we want the block post to contain the title the publish date the content and mind that this is always

serialized by default only because we haven't previously provided that for certain action.

We would like our specific group to be used for that normalization.

And since we defined that for item operation get we would like our specific group.

Now if you do that now for those cases you always have to the group annotation on the fields that you'd

like to serialize.

Actually we don't have to add it to comments because we have this sub resource you are under which we

can fetch comments.

But let's just add it and see what happens.

Probably nothing wrong but let's just put up now inside the user.

You'll have to do the same.

What we did with their comments list so we will adding We are adding a new group on each of their fields

that we want to serialize.

When we are showing an individual block posts so we definitely do not want this on posts or comments

because we only need the username and the full name of the user.

So after refreshing you can see that we have the data of the blog post.

Everything that we've chosen to serialize and we also have a nested outer entity with the user name

and the form name.

We also have a list of comments we have there you are ice.

Now let's verify if we fetch a list of all the blog posts that outer Issiah you or I.

And it is.

So the change that we made only applies to an individual blog posts and we are not fetching it for a

list of blog posts.

If you'd like to then you should change that get operation for collection operations to have a normalization

context and a proper group said.

We have also not broken anything for their comments sublease.

It still works as expected.