83. Configuring custom operation for password reset in User 84.1 Creating custom Action class(Section 12: API Platform - Custom Operations (Password Reset))
Create custom password reset operation

84. Creating custom Action class

In the last lecture we defined our custom operation and forward the custom operation.

We actually have to implement them manually.

So we said that the controller for this operation would be reset password action and API platform requirements

that you would you should create so-called auction classes for the custom operations instead of creating

fully fledged controllers like you would do inside scene from them.

So since we have added it to the controller namespace it can be automatically out of wired by Sinfonie

because this namespace is important.

So everything inside the controller will be out to wire.

And now let's mettled called invoke.

This is special method from the BHB So each time you will define a method invoke when you will create

a new instance of that class then you could call it like a metal.

So like DUP.

So if you will create a new instance of new reset reset password action then you can just call it as

method.

So the whole class is like a single metal that you will.

You can call after creating this class

so we can add parameter called user because we will be working on that user entity this operation custom

operation of user.

So Sinfonia so API platform actually will pass instance of the user that it does serialize the serialise

from the data that we are sending

So lets first outing teacake as ropes Meave will be trying to modify ropes me of password reset password

so lets use that default password that he has currently from the fixtures.

That's why we reloaded the fixtures some time ago and that idea of ropes Meave is 81.

So currently this 87 our data was out of sync.

We haven't refreshed it after loading fixtures though so be sure to check yours.

So you get the proper I.D. and let's type.

There you are.

I'll forward that action inside postman and be sure to choose the port a verb us HTP method.

Now let's try to send it.

So we provided new password new Rito password but we have an error that the class does not exist and.

Well it happens sometimes our forgot to import this class inside the user entity.

So the annotation could not figure out what the class name of that entity because it needs to be imported.

So you need to add the use statement now actually we try to damp the whole data.

And for some reason this might be a pretty big chunk of data.

So the postman is having problems rendering it so maybe let's modify this statement because we are first

trying to see what will be sent here.

So let's try to just dump the individual properties that we are expecting to receive inside this action

which is the new password new retyped password and that old password and after that we are just dying.

Because we just want to see what are we getting.

You know so you can see that we have new password new retired password and old password.

Exactly.

We have sent it now you might expect that the validation will happen before we get this data passed

to our customer action.

But actually it will happen after we return that data from our custom operation.

And if we will and coat the passport so we will set the password of the user and let it then validate

and persist our anti-Tea modified entity which happens automatically in site.

Even the custom operations the validation for the old password will fail because it will try to fetch

the password from their database.

And we have already modified it.

And the way doctrine works is even if it has already the copy of entity loaded from the memory it will

just use it so it will not fetch it from the database.

So to solve that we will have to manually validate our data inside our action.

And this might be useful because you will see how you can perform all this all the steps yourself.

So we are writing on top early custom operation and we'll do everything manually.

So you can not validate that data.

If your custom operation does something that is not a typical data that you send and store but it is

something that doesn't even have an effect or is sending some e-mails.

So now you can see that we have a problem.

So we are validating the data and the validator is showing us errors that they're retyped.

Password is blank and that it does not match that original password.

So we we didn't expect to file a different property.

So we didn't expect it to validate retired passport because we are sending different fields.

So let's see what's the value of the retired password.

And you can see that it is no.

So the validator is right this this field should not be empty but we do not want this field to be checked.

The constraints of this field to be checked inside this particular action and the way we have our validation

constraints defined is that they will work always or older data that we send.

So for all the actions for all the post and port operations so we can also use groups for the validator

and this will be the same group that you defined for either the normalization or in this case it is

a normalization process the writing process.