84.2 Creating custom Action class(Section 12: API Platform - Custom Operations (Password Reset))
So we send the data and by adding groups property to each of the validation constraints in each of the

properties we make it so that it will only use the validator we only check for this contract constraints

on these properties if the current this realisation context the group is the same as you define it inside

the constraints so we only want those things to be checked when the group is post or in some cases we

also want it to be checked when the group is put.

So we allow users to modify their name.

So when this data is being sent through the normal post and put operations we want it to be verified

or maybe accept we do not want to verify if for example name is not blank.

Inside the pute operation because boot operation does not require all the parameters to be present only

those we that we would like to modify.

So by setting the not blank constrained groups to post when user registers We require all the fields

to be provided input operation.

We only require the very patient constraints that will actually check the data.

Like lenth will only check the data is there.

If the value is provided to all so by having the land property it does not require that value to be

present at all.

If it's not present it just won't be checked.

So now let's see again what the validator will say.

So it did.

Got that right.

New passport and you retyped passport are indeed different.

You can see that there is one L letter missing.

And also if you modify the old passport to the different one that is being stored you can see that it

also is saying that the value should be the user's current password.

So that's it for now.

We'll continue with this in the next lecture.