85.1 Implementing custom PasswordReset action(Section 12: API Platform - Custom Operations (Password Reset))
So let's go back to our custom operation and try to provide data that will pass the validation.

So the passports should match.

And you can see that we have a different error that it cannot validate values of type.

No but we have just called the validator.

So let's see why does that happens.

So what it actually expects is that you will do something with that entity and then you will return

it from this custom operation and then API platform will validate and persist this anti-Tea.

And you can also see that it actually returned on e-mail and rolls.

And that's why because we haven't defined normalization context for user the default one.

So since we are authenticated us ourselves and thanks to our custom normalizer it adds certain groups

to the serializer if the user is actually himself.

So we might fix this later.

But in general we would like to override this default behavior of API platform and now we validate the

data.

And this is fine but then if we would actually modify the password and return that data back to the

API platform to handle it further.

It will again try to validate it and persisted and this wouldn't work because the user password validator

will stop the violation because we are changing the password right here and coding it.

So you can see that it is complaining that the old password is incorrect because we have just modified

it.

So that's what I mentioned in the last lecture that if we modify that entity and it is being loaded

already in inside the doctoring unit of work.

So it tries to just optimize the way it handles the database so if there is a call to load the entity

from the database and the risk to check the current password it would just get that entity that is already

loaded.

And on this entity we have modified the password.

So this just would not work so we cannot delegate the validation and saving to API platform.