85.2 Implementing custom PasswordReset action(Section 12: API Platform - Custom Operations (Password Reset))
We all have to actually handle this inside our custom operation.

So we are defining this custom operation just as an example but so you would know that each time you

will have your custom operations and you will also like to use some of the API platform features like

for example validating for you ultimately and persisting.

You have to be aware that if you're just modify something and then validate the way that those modifications

will not let the validation pass.

You might have problems like that.

So now instead of returning the entity from this custom operation we would just persisted ourselves

because we are performing the validations before we modify the password to the new one.

So if someone is sending the old password as a proper one then divide nation will pass and only then

we will automatically we will persist explicitly there.

And just to this anti-Tea

You can also see that we are calling on the flash we do not call the persist method and this is because

we only call Percy's for the new entities if we just created them and we want to insert them to the

database we call persist.

But when we have an existing entity that we modify it then we do not have to call Percy's because doctorin

will ultimately figure out what has changed.

And just by calling flash it will execute the proper update execute statements.

OK so let's go back to the postman and we have modified the password for the ropes move.

So it's not that secret want to free Hosh anymore like we have defined inside the fixtures.

It is being modified to the value that we have sent inside our reset password auction.

So that's not what we expected because our auction is not done yet but it was modified.

So now let's try to log in with this new password and let's get the talken and set it to our reset password.

Auction

So the old password is now not secret one to free house but it is actually.

Hello.

One two three dollar and a new password.

Let's maybe change it again to the password that we had before.

So at this point we are saving that changes to the database.

But it might be useful if the user is changing password to automatically generate new talken for that

user so he won't have to log in again.

We can just automatically keep him sending requests using there all the requests for any other resources

using their newly generated talkin.

And by the way you can see how you can generate the and programmatically.

So it's just enough to import those JPT class that we added to the constructor.

It will be ultimately injected and if you will pass a user instance to the doc and manager create it

will generate new token strings so we can then return it from this from this custom action.

And while we have the new password and our constraints to not allow this because we need to have an

upper case so that's what we have defined for passwords and the fixtures are not being checked.

So that's why we had different password.

Now you can see that by changing the password for this particular user.

Now it worked.

We have generated new token for that user so now to actually logging We can either try logging again

or we can just start using this new token that we got back from our reset password.

Auction.

So obviously after changing the passwords our old password stopped working so we were able to authenticate.

But I think this is quite a good solution to return the new token if the password has changed.