86.1 Invalidating JWT tokens after password reset(Section 12: API Platform - Custom Operations (Password Reset))
So last time we created our custom operation that lets users reset their password modify it to a different

one and then YSU returning our brand new JPT token for that user.

So we've also seen how simply we can generate new tokens programmatically outside the built in functionality

that the JOAT bundle provided to us.

But the problem is that after we have changed the password the old tokens are still valid because Talkin

is a pretty simple thing.

It just contains a user name and the expiry date which is signed by our keys.

So it is always valid because that's only thing that is being checked inside that token authenticator.

So they get credentials is trying to decode the token and it is creating an instance of pre token which

are the user credentials.

It's the class that TVT Bondel is holding the credentials inside.

And after decoding this token it just tries to load the user by the user name that was stored inside

the token.

So the only security that we have is that we are checking using the private public key cryptography

that we we ourselves our application have created that token.

So that's the only security check we do.

So all the tokens that were generated and are within the 1 hour time to leave will remain to be valid

even if you have changed password.

And probably they shouldn't be.

So we will handle that case right now by adding a new column to the user entity which will hold the

password change day.

So whenever you have changed your password it will hold the date and it is an integer not a date time

because we will keep our Unix time stamp inside it.

Not the usual date time and that is because the JOAT token inside the JVC token that we generate we

keep the date of when the token will expire.

Our Unix timestamp so it is on integer of the seconds that have passed.

Seems the Unix epoch which is nineteen seventy.