86.2 Invalidating JWT tokens after password reset(Section 12: API Platform - Custom Operations (Password Reset))
So after we changed the password we will also store it inside the database and it will always store

the date of the last password change and we can again rate Unix timestamp by calling their pute B time

function.

So now that we are storing the time of the password change we will have to also create our custom out

indicator class.

So that's the that would do the same as I showed you in that lectures about security that built in JPT

talkin out in Decatur which is our guard out in Decatur which uses the symphony guard system for easily

implementing the out indication.

Now we won't overwrite much of the logic from this.

We will only override the get user function.

So we are extending that GJ Viti token authenticator and we will be calling its parent method to get

user because that logic will be useful to us and this method should return user if it can fetch any.

So if this method returns a user instance there Sinfonie will treat the user as properly authenticated.

So here we will out our checks so we get the user provider in user provider interface passed to this

method.

And this is the anti-Tea user provider that we have configured in lectures about security so we don't

really have to worry about that.

We can just use its made up load user by username.

OK so now let's see what do we have from the get credentials mettled which is being called just before

they get to user by Sinfonia and their credentials that it was able to decode from that token are being

passed inside this prie our token class.

So first let execute them migration.

And since we made that field new label it is safe to be no by default because not everyone will change

their passwords so it could be no and only after someone changes the password.

It has to be modified.

Now let's try to log in as that one.

So for example let's switch the users don't bother the one user with all these changes all the time.

Next let's try to call the reset password operation using using the token that we have just got from

logging in

and Han Solo.

Current password is secret.

Want to rehash and we will try to modify this password.

So let's see what will happen.

OK so we have a new new token.

And we successfully modified the user password but also we have stored the password change date.

So this is the Unix time timestamp of when the password of that user was actually changed.

So now we have our token authenticator that will take over the authentication process after you.

If you are registered and then you will like to authenticate other requests with the token that you

have fetched from the logon page.

Now let's first see what's the class behind this service that is defined to you.

So this is the original lexia Jay Viti talkin authenticator.

And since we have overridden that class we can use our class instead of the original from the next week.

Viti bandeau and our token out in Decatur will be out all wired so we can get this we can get that functionality.

Now we can just call the reset password just because we have a very token sent to it and you can see

that they there are a couple of there's some information stored inside the payload of the token.

So what we saw there was the token that was coded by the previous metadata that is called just before

that user they get credentials.

Mettled.

Now let's also add a page be doc to to have the code completion inside our ID and by adding this var

PH BE doc.

That future system will be able to know what's the type of that object that will return.

Now we have couple of properties here.

There are the roles of the user the user name ID and expr.

So we can use that.

I see property which is the date where the token was generated so we can compare that to our password

change date of the user.

And if that token was generated before our password was changed we will actually follow an exception

here.

So we can use one of exception classes that we already have expired doc an exception and this makes

sense because talken is expired actually.

And only if that token was generated before we can change the password for all these exceptions.

And in all other cases we we just process as usual so we just return this user because that means that

he was properly authenticated.

So you can see that that token was generated before we have changed the password.

So you can see that we use that token that we have fetched before and now we have modified the user.

So we when we called this action again it just didn't let it go.

So this is a pretty useful thing and I think that you should also implemented inside your API some way

of checking if tokens are still valid and definitely if after user has changed his password maybe he

changed the password because someone tried to compromise his account.

So it's a pretty good idea to always reset all the old tokens after a password change too.

Basically it's for the security of your own users.